i18n fetch: Reduce number of requests to MW API
Instead of downloading the MW messages specified in twinkle-core and the ones specified in the twinkle client separately, merge the arrays and request them together. `Twinkle.extraMwMessages` has been created for wikis to specify the MW messages they need. This also slightly simplifies the code needed in the client's twinkle.ts file.