i18n: unbreak existing release of twinkle-core
I didn't realise earlier that using `fallbacks` as a key in the built i18n files (which are deployed immediately) would cause issues in the existing twinkle-core release which doesn't have the code to delete the property before passing to `banana.load`. Using `@fallbacks` instead fixes this as banana ignore props beginning with `@`.